# Cache HW device context