--- a/torchaudio/csrc/ffmpeg/hw_context.h
+++ b/torchaudio/csrc/ffmpeg/hw_context.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <torchaudio/csrc/ffmpeg/ffmpeg.h>
+
+namespace torchaudio::io {
+
+AVBufferRef* get_cuda_context(int index);
+
+void clear_cuda_context_cache();
+
+} // namespace torchaudio::io